Tidy up and simplification
